+++ src/index.jsx
@@ -1,69 +1,6 @@
-import React, { useState } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
-
-function Button(props) {
-  const { children } = props;
-  return (
-    <button type="button">
-      {children}
-    </button>
-
-  );
-}
-
-function Buttons() {
-  return (
-    <p>
-      {[1, 2, 3, 4, 5].map((i) => (
-        <Button key={i}>
-          {i}
-        </Button>
-      ))}
-    </p>
-  );
-}
-
-function Counter({ count, onClick }) {
-  return (
-    <button type="button" onClick={onClick}>
-      Click me! (
-      { count }
-      )
-    </button>
-  );
-}
-
-function Page({ count, onClick }) {
-  return (
-    <div>
-      <p>Counter</p>
-      <Counter
-        count={count}
-        onClick={onClick}
-      />
-      <Buttons />
-    </div>
-  );
-}
-
-function App() {
-  const [state, setState] = useState({
-    count: 0,
-  });
-  const { count } = state;
-
-  function handleClick() {
-    setState({
-      count: count + 1,
-    });
-  }
-  return (
-    <Page
-      count={count}
-      onClick={handleClick}
-    />
-  );
-}
+import App from './components/App.jsx';
 
 ReactDOM.render(
   <App />,
